requirements.txt
openpyxl>=3.1.5
Pillow>=11.2.1
segno>=1.6.6
tqdm>=4.67.1